Add support for manual cheat creation

+++ app/src/main/java/me/magnum/melonds/impl/RoomCheatsRepository.kt
@@ -9,6 +9,9 @@ import androidx.work.WorkInfo
 import androidx.work.WorkManager
 import androidx.work.workDataOf
 import io.reactivex.Observable
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.emptyFlow
+import kotlinx.coroutines.flow.map
 import me.magnum.melonds.common.workers.CheatImportWorker
 import me.magnum.melonds.database.MelonDatabase
 import me.magnum.melonds.database.entities.CheatDatabaseEntity
@@ -23,6 +26,7 @@ import me.magnum.melonds.domain.model.CheatImportProgress
 import me.magnum.melonds.domain.model.Game
 import me.magnum.melonds.domain.model.RomInfo
 import me.magnum.melonds.domain.repositories.CheatsRepository
+import me.magnum.melonds.ui.cheats.model.CheatSubmissionForm
 
 class RoomCheatsRepository(private val context: Context, private val database: MelonDatabase) : CheatsRepository {
     companion object {
@@ -53,23 +57,41 @@ class RoomCheatsRepository(private val context: Context, private val database: M
         }
     }
 
-    override suspend fun getAllGameCheats(game: Game): List<CheatFolder> {
-        val gameId = game.id ?: return emptyList()
+    override fun getAllGameCheats(game: Game): Flow<List<CheatFolder>> {
+        val gameId = game.id ?: return emptyFlow()
 
-        return database.gameDao().getGameCheats(gameId).map {
-            CheatFolder(
-                it.cheatFolder.id,
-                it.cheatFolder.name,
-                it.cheats.map { cheat ->
-                    Cheat(
-                        cheat.id,
-                        cheat.name,
-                        cheat.description,
-                        cheat.code,
-                        cheat.enabled
-                    )
-                }
-            )
+        return database.gameDao().getGameCheats(gameId).map { foldersWithCheats ->
+            foldersWithCheats.map {
+                CheatFolder(
+                    it.cheatFolder.id,
+                    it.cheatFolder.name,
+                    it.cheats.map { cheat ->
+                        Cheat(
+                            cheat.id,
+                            cheat.cheatDatabaseId,
+                            cheat.name,
+                            cheat.description,
+                            cheat.code,
+                            cheat.enabled
+                        )
+                    }
+                )
+            }
+        }
+    }
+
+    override fun getFolderCheats(folder: CheatFolder): Flow<List<Cheat>> {
+        return database.cheatDao().getFolderCheats(folder.id!!).map {
+            it.map { cheat ->
+                Cheat(
+                    cheat.id,
+                    cheat.cheatDatabaseId,
+                    cheat.name,
+                    cheat.description,
+                    cheat.code,
+                    cheat.enabled
+                )
+            }
         }
     }
 
@@ -77,6 +99,7 @@ class RoomCheatsRepository(private val context: Context, private val database: M
         return database.cheatDao().getEnabledRomCheats(romInfo.gameCode, romInfo.headerChecksumString()).map { cheat ->
             Cheat(
                 cheat.id,
+                cheat.cheatDatabaseId,
                 cheat.name,
                 cheat.description,
                 cheat.code,
@@ -93,11 +116,36 @@ class RoomCheatsRepository(private val context: Context, private val database: M
         database.cheatDao().updateCheatsStatus(cheatEntities)
     }
 
-    override fun deleteCheatDatabaseIfExists(databaseName: String) {
-        database.cheatDatabaseDao().deleteCheatDatabase(databaseName)
+    override suspend fun addCheatFolder(folderName: String, game: Game) {
+        val gameId = if (game.id == null) {
+            // It's a new game. Insert it first
+            val gameEntity = GameEntity(
+                id = null,
+                name = game.name,
+                gameCode = game.gameCode,
+                gameChecksum = game.gameChecksum
+            )
+            database.gameDao().insertGame(gameEntity)
+        } else {
+            game.id
+        }
+
+        val cheatFolderEntity = CheatFolderEntity(null, gameId, folderName)
+        database.cheatFolderDao().insertCheatFolder(cheatFolderEntity)
     }
 
-    override fun addCheatDatabase(databaseName: String): CheatDatabase {
+    override suspend fun deleteCheatDatabaseIfExists(databaseName: String) {
+        if (databaseName == CheatDatabaseEntity.CUSTOM_CHEATS_DATABASE_NAME) {
+            // Don't allow the custom cheat database to be deleted
+            return
+        }
+
+        database.cheatDatabaseDao().deleteCheatDatabase(databaseName)
+        database.cheatFolderDao().deleteEmptyFolders()
+        database.gameDao().deleteEmptyGames()
+    }
+
+    override suspend fun addCheatDatabase(databaseName: String): CheatDatabase {
         val cheatDatabaseEntity = CheatDatabaseEntity(
             null,
             databaseName,
@@ -107,21 +155,24 @@ class RoomCheatsRepository(private val context: Context, private val database: M
         return CheatDatabase(databaseId, databaseName)
     }
 
-    override fun addGameCheats(databaseId: Long, game: Game) {
+    override suspend fun addGameCheats(game: Game): Game {
         val gameEntity = GameEntity(
-                null,
-                databaseId,
-                game.name,
-                game.gameCode,
-                game.gameChecksum
+            null,
+            game.name,
+            game.gameCode,
+            game.gameChecksum
         )
 
-        val gameId = database.gameDao().insertGame(gameEntity)
+        // Insertion may do nothing if the game already exists
+        database.gameDao().insertGame(gameEntity)
+        val insertedGame = database.gameDao().findGame(game.gameCode, game.gameChecksum)!!
+        val gameId = insertedGame.id!!
+
         val categoryEntities = game.cheats.map { category ->
             CheatFolderEntity(
-                    null,
-                    gameId,
-                    category.name
+                null,
+                gameId,
+                category.name
             )
         }
         val categoryIds = database.cheatFolderDao().insertCheatFolders(categoryEntities)
@@ -129,16 +180,73 @@ class RoomCheatsRepository(private val context: Context, private val database: M
         val cheatEntities = game.cheats.zip(categoryIds).flatMap { pair ->
             pair.first.cheats.map {
                 CheatEntity(
-                        null,
-                        pair.second,
-                        it.name,
-                        it.description,
-                        it.code,
-                        false
+                    id = null,
+                    cheatFolderId = pair.second,
+                    cheatDatabaseId = it.cheatDatabaseId,
+                    name = it.name,
+                    description = it.description,
+                    code = it.code,
+                    enabled = false
                 )
             }
         }
         database.cheatDao().insertCheats(cheatEntities)
+
+        return Game(
+            id = insertedGame.id,
+            name = insertedGame.name,
+            gameCode = insertedGame.gameCode,
+            gameChecksum = insertedGame.gameChecksum,
+            cheats = emptyList(),
+        )
+    }
+
+    override suspend fun addCheat(folder: CheatFolder, cheat: Cheat) {
+        val cheatEntity = CheatEntity(
+            id = null,
+            cheatFolderId = folder.id!!,
+            cheatDatabaseId = cheat.cheatDatabaseId,
+            name = cheat.name,
+            description = cheat.description,
+            code = cheat.code,
+            enabled = cheat.enabled,
+        )
+
+        database.cheatDao().insertCheat(cheatEntity)
+    }
+
+    override suspend fun addCustomCheat(folder: CheatFolder, cheatForm: CheatSubmissionForm) {
+        val cheatEntity = CheatEntity(
+            id = null,
+            cheatFolderId = folder.id!!,
+            cheatDatabaseId = CheatDatabaseEntity.CUSTOM_CHEATS_DATABASE_ID,
+            name = cheatForm.name,
+            description = cheatForm.description,
+            code = cheatForm.code,
+            enabled = false,
+        )
+
+        database.cheatDao().insertCheat(cheatEntity)
+    }
+
+    override suspend fun updateCheat(cheat: Cheat) {
+        val originalEntity = database.cheatDao().getCheat(cheat.id!!) ?: return
+        val updatedCheatEntity = CheatEntity(
+            id = cheat.id,
+            cheatFolderId = originalEntity.cheatFolderId,
+            cheatDatabaseId = originalEntity.cheatDatabaseId,
+            name = cheat.name,
+            description = cheat.description,
+            code = cheat.code,
+            enabled = cheat.enabled,
+        )
+
+        database.cheatDao().insertCheat(updatedCheatEntity)
+    }
+
+    override suspend fun deleteCheat(cheat: Cheat) {
+        val cheatId = cheat.id ?: return
+        database.cheatDao().deleteCheat(cheatId)
     }
 
     override fun importCheats(uri: Uri) {
